refactor: 문장 목록을 반환하는 유틸 적용
